Remove warning by default
I think this log does make a lot of sense if one is a Jenkins developer
checking what classes/plugins need fixing.

But this warning is probably not actionable by end users, and this is
generating a lot of logs with current Jenkins cores and plugins available
given a lot of fixes have yet to land.

So I'm lowering this so that it's still present, but one (developer) can
enable it if needed.
I think this will probably be fine to put it back to INFO or WARNING much
later when we think we have fixed everything and want to watch the
latest remaining bad descriptors.