Add support for bcrypt hashed passwords
I'm unsure if this was the best approach to solving this problem.
The main issue is that most of the code for constructing
hashed user credentials in HudsonPrivateSecurityRealm is private
or package private, and thus, we can't just pull it in.

So I'm taking a reflective approach for this, with the eventual
hope to add an upstream feature to Jenkins to support this. However,
that would cause an API break and mandatory Jenkins upgrade for
all users of this plugin, which is a fairly comprehensive change.

I believe the unit test coverage for this should hopefully cover
method name breaks in Jenkins core.