Added escape for secrets to docs
Per #697 proper interpolation was added. However, some plugins such as the kubernetes plugin require ${computer.jnlpmac} and ${computer.name} literals in the config.xml. The PR didn't add this extra bit of info for escaping the secrets and would require someone to review the PR to see the escape character of `^`.