Move ConfigurationContext one level up to enable configurable max aliases config
This needs to be a property/env since the context is not yet configurable as it has yet to read any YAML
and it might be blocked due to alias limit.

Adding the reading of the property/env to ConfigurationContext
makes sense to avoid YamlUtils to constantly reading the property on every file read
when it attempts to merge all the sources.