Add a note on reloading from declarative pipelines #1227
If JCasC is reloaded from a declarative pipeline, Jenkins throws an error due to permission issues. Subsequent reloads of JCasC fail regardless of the mechanisms used to do so, and regardless of user permissions.

This issue probably needs to be addressed correctly so that Jenkins isn't put in an invalid state, as it could potentially have other side effects that hasn't been noticed yet.