Update Git plugin and Git client plugin examples
The git client plugin example was incorrectly placed in a directory
named "git" while the git plugin example was placed in a directory named
"gitscm".  Let's use directory names that match the plugin identifier
as is used for the other plugin examples in the demos directory.

Add ssh host key verification policy configuration to the git client
plugin example and include a link to the plugin documentation.

Add more git plugin configuration elemnts and link to the git plugin
documentation for more details.

Fixes #2410