Run only 80% of Windows integration tests on ci.jenkins.io
Reduce the risk of test timeout on ci.jenkins.io and reduce the cost of
test execution.

Rotates which 80% of tests are executed based on the build number
so that every test will be executed 4 times in a series of 5 runs on
ci.jenkins.io.

If a test is skipped and that hides a failure, the next run of the
ci.jenkins.io job will show the failure.