Remove Crowd2 demo and integration test - suspended plugin
https://github.com/jenkins-infra/helpdesk/issues/3854 notes that
the Crowd2 plugin uses one or more closed source dependencies.
Those dependencies make it ineligible to be distributed by the Jenkins
update center.  Let's remove the integration test and documentation so
that we are not describing the configuration of a suspended plugin.

https://github.com/jenkins-infra/helpdesk/issues/3842#issuecomment-1839385370
describes the case where we detected the closed source dependency
through the configuration as code plugin integration test of the Crowd2
plugin.  Unless we include the Atlassian closed source repository,
the configuration as code plugin fails to compile its integration tests.

https://github.com/jenkins-infra/helpdesk/issues/3842#issuecomment-1841501804
provides additional details from the investigation related to the Jenkins
artifact repository and its caches.

https://github.com/jenkins-infra/helpdesk/issues/3842#issuecomment-1847559631
summarizes my investigation