Require Jenkins 2.426.3 or newer
https://stats.jenkins.io/pluginversions/configuration-as-code.html shows
that 93% of installations of the most recent release of configuration
as code plugin are already running Jenkins 2.426.3 or newer.

https://www.jenkins.io/security/advisory/2024-01-24/#SECURITY-3314
advises administrators to install Jenkins 2.426.3 or newer to resolve
that critical security vulnerability.

https://stats.jenkins.io/pluginversions/configuration-as-code.html shows
that 50% of all installations of configuration as code plugin are already
running Jenkins 2.426.3 or newer.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
recommends Jenkins 2.426.3 or Jenknis 2.440.3 as recommended base version.